+++ src/oersted/magnetization.py
@@ -9,7 +9,7 @@ from numpy import float64, uint32
 from .mesh import Mesh
 from .constants import MU0
 from .materials import Material, LinearMaterial
-from ._oersted import _h_demag_tet4, _hfield_dipole_tetrahedrons
+from ._oersted import h_mag_tet4_direct, _hfield_dipole_tetrahedrons
 
 
 def mag_force(centroids: NDArray[float64], vol: NDArray[float64], material: LinearMaterial, h_ext: Callable) -> NDArray[float64]:
@@ -66,12 +66,10 @@ def mag_force(centroids: NDArray[float64], vol: NDArray[float64], material: Line
 
 
 def h_demag_tet4(
-    src_nodes: NDArray[float64],
-    src_connectivity: NDArray[uint32],
+    mesh: Mesh,
     material: Material,
     m_field: NDArray[float64],
-    tgt_nodes: NDArray[float64] | None = None,
-    tgt_connectivity: NDArray[uint32] | None = None,
+    targets: NDArray[float64] | None = None,
     nthreads_requested: int = 0,
 ) -> NDArray[float64]:
     """Compute the demagnetization field H(M) on mesh element centroids given the current M-field
@@ -91,35 +89,20 @@ def h_demag_tet4(
 
     # Check that the M field is calculated at the element centroids
     try:
-        assert src_connectivity.shape[0] == m_field.shape[0]
+        assert mesh.connectivity.shape[0] == m_field.shape[0]
     except AssertionError:
         print("Error. The M-field should be calculated at element centroids.")
 
-    if tgt_connectivity is None:
-        tgt_connectivity = src_connectivity
+    if targets is None:
+        targets = mesh.centroids
 
-    if tgt_nodes is None:
-        tgt_nodes = src_nodes
-    n_elements: int = tgt_connectivity.shape[0]
-    hx = np.zeros(n_elements)
-    hy = np.zeros(n_elements)
-    hz = np.zeros(n_elements)
-
-    _h_demag_tet4(
-        np.ascontiguousarray(src_nodes.flatten()),
-        np.ascontiguousarray(src_connectivity.flatten().astype(np.uint32)),
-        np.ascontiguousarray(tgt_nodes.flatten()),
-        np.ascontiguousarray(tgt_connectivity.flatten().astype(np.uint32)),
-        np.ascontiguousarray(m_field[:, 0]),
-        np.ascontiguousarray(m_field[:, 1], dtype=float64),
-        np.ascontiguousarray(m_field[:, 2]),
-        np.ascontiguousarray(hx),
-        np.ascontiguousarray(hy),
-        np.ascontiguousarray(hz),
+    return h_mag_tet4_direct(
+        np.ascontiguousarray(mesh.nodes),
+        np.ascontiguousarray(mesh.connectivity),
+        np.ascontiguousarray(m_field),
+        np.ascontiguousarray(targets),
         nthreads_requested,
     )
-
-    return np.hstack((hx[:, np.newaxis], hy[:, np.newaxis], hz[:, np.newaxis]))
 
 
 def h_demag_tet4_octree(
@@ -229,7 +212,9 @@ def demag_tet4(
             h_demag = h_demag_tet4_octree(mesh.nodes, mesh.connectivity, material, m_field, centroids, vol, nthreads_requested=nthreads_requested)
 
         else:
-            h_demag = h_demag_tet4(mesh.nodes, mesh.connectivity, material, m_field, nthreads_requested=nthreads_requested)
+            h_demag = h_demag_tet4(mesh, material, m_field, nthreads_requested=nthreads_requested)
+
+        h_demag = h_demag_tet4(mesh, material, m_field, nthreads_requested=nthreads_requested)
         h_total = h_demag + h_external
 
         # We consider isotropic materials for the B-H curve iteration