+++ CHANGELOG.md
@@ -64,3 +64,5 @@
 * Updated and simplified helmholtz coil and solenoid tests; added analytical comparisons
 * Cleanup duplicated functions
 * Move `Vec3` and Mat3 into `types`
+* Updated magnetized tetrahedral calcs to use analytical gradient and significantly cleaned up related functions
+* Added `SurfaceMesh` class